Close linux-build-pthread-mutex.sh's own compile+link gaps
Never actually compiled before on any arch: CI's `bash -e` loop always
stopped at an earlier script (tools/linux-build-crt.sh, then
tools/linux-build-fs.sh) until this session fixed them, so nothing
reached this one.

Two real, independent bugs surfaced:

  - fuzz/linux_pilot_harness_pthread_mutex.c's calloc() stub declared
    `unsigned long nmemb, unsigned long size`, but this tree's size_t
    (obj/include/bits/alltypes.h's _Addr) is `long long` on this arch --
    a distinct type at the same width, so clang rejected it as a
    conflicting redeclaration against <stdlib.h>'s real prototype.
    Fixed by matching the real signature (size_t).
  - src/unistd/linux/plat_unistd.c's syncfs() calls the real
    __fd_get(), missing src/internal/fd.c + src/internal/linux/
    plat_fd_init.c from this script's own FILES list. Unlike
    tools/linux-build-fs.sh/-open.sh/-misc.sh's own __free gap, nothing
    reachable here calls __fd_install_at(), so --gc-sections drops
    __fd_release_dynamic()'s own __free() call before the link needs
    it -- no allocator files were needed here.

Verified with a full clean rebuild+run: linux-build-pthread-mutex
reports its own "ALL CHECKS PASSED" / "linux-build-pthread-mutex: PASS"
across both its test groups (basic mutex lifecycle,
PTHREAD_MUTEX_RECURSIVE).

+++ fuzz/linux_pilot_harness_pthread_mutex.c
@@ -51,9 +51,9 @@ void __sig_current_mask_copy(sigset_t *mask)
 static unsigned char pool[POOL_SLOTS][256];
 static int pool_next;
 
-void *calloc(unsigned long nmemb, unsigned long size)
+void *calloc(size_t nmemb, size_t size)
 {
-	unsigned long total = nmemb * size;
+	size_t total = nmemb * size;
 	int i, slot;
 	if (total > sizeof pool[0]) return 0;
 	/* Real concurrent callers: every worker thread's first
